Refactor to use generic views 💥
Django provides abstractions we can use for these concepts:

* `ListView` to display a list of objects
* `DetailView` to display a detail page for a particular object

The `question` context variable can be deleted now, because using the
model `Question` means Django determines it automatically for detail
and result. But for index using `ListView` we want to override the
context variable with a `context_object_name` attribute. Otherwise we
would get the default `question_list` instead of `latest_question_list`.